requirements.txt
streamlit
tavily-python
google-genai
ollama
python-dotenv
requests
pytest
